--- src/main/java/com/project/controllers/FooterController.java
@@ -1,22 +1,0 @@
-package com.project.controllers;
-
-
-import org.springframework.stereotype.Controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-
-
-
-    @Controller
-    public class FooterController {
-
-
-
-
-
-
-        @GetMapping("/inter")
-        public String getInternationalPage() {
-            return "inter";
-        }
-    }
